pyproject.toml
[tool.poetry]
name = "backend"
version = "0.0.1"
description = "A fastAPI service for GVSC DSDP project"
authors = ["Anthony Butt <anthony.butt@cse-corp.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
python-multipart = "^0.0.5"
uvicorn = "^0.17.6"
PyJWT = "^2.4.0"
PyYAML = "^6.0"
cryptography = "^37.0.2"
boto3 = "^1.23.8"
Secweb = "^1.5.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"
pytest-cov = "^3.0.0"
pytest-mock = "^3.7.0"
pylint = "^2.13.9"
requests = "^2.27.1"
pytest-dotenv = "^0.5.2"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov --cov-config=pyproject.toml --cov-report=term-missing"

[tool.coverage.run]
omit = ["tests/*",".cache/python-packages/*"]

[tool.pylint.master]
disable = ["C0415", "E0401"]
extension-pkg-whitelist = ["pydantic"]

[tool.pylint.design]
min-public-methods = 0

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
